refactor: split monolithic load_fixtures into modular fixture modules and add --output-json
Break the 1,751-line load_fixtures.py into focused modules under fixtures/:
- __init__.py: shared utilities (faker setup, phone generator, valueset bypass)
- organizations.py: geo/role/facility organization factories
- facilities.py: facility, location, and device factories
- users.py: user creation with role attachment
- patients.py: patient and encounter factories
- clinical.py: lab definitions (specimen, observation, activity definitions)
- inventory.py: products, inventory items, supply deliveries
- questionnaires.py: questionnaire loading from JSON

The main load_fixtures.py is now a slim orchestrator (~220 lines) that imports
and calls functions from these modules.

Add --output-json flag that writes a JSON manifest of all created entity IDs
(facility, patients, encounters, users, locations, devices). This enables
frontend Playwright tests to read IDs directly instead of navigating the UI
to extract them.
